add net config

+++ scripts/run-vm-vps.sh
@@ -2,17 +2,33 @@
 # qemu-img create -f qcow2 -F qcow2 -b arch.qcow2 arch-vm1.qcow2
 set -euo pipefail
 
-VM_ID="${VM_ID:-1}"
-VM_IMG="${VM_IMG:-images/linux-$VM_ID.qcow2}"
+VM_IMG="${1:-images/linux-$VM_ID.qcow2}"
+VM_ID="${2:-1}"
 VM_PORT=$(printf '%02u' "$VM_ID")
 MAC_SUFFIX=$(printf '%02x' "$VM_ID")
 VM_MAC="52:54:00:12:34:${MAC_SUFFIX}"
 TAP_IF="tap-k8s${VM_ID}"
 
+if [[ ! "$VM_IMG" =~ ^/ ]]; then
+    VM_IMG="$(realpath "$VM_IMG")"
+fi
+
+# Check if image exists
+if [[ ! -f "$VM_IMG" ]]; then
+    echo "Error: Image file not found: $VM_IMG"
+    exit 1
+fi
+
+PROJECT_ROOT=$(dirname "$(dirname "$VM_IMG")")
+VM_NAME=$(basename "$VM_IMG" .qcow2)
+OVMF_VARS_DIR="${PROJECT_ROOT}/ovmf_vars"
+OVMF_VARS="${OVMF_VARS_DIR}/OVMF_VARS_${VM_NAME}.4m.fd"
 OVMF_CODE=/usr/share/OVMF/x64/OVMF_CODE.4m.fd
-OVMF_VARS=$HOME/workspace/vpc-qemu/OVMF_VARS_${VM_ID}.4m.fd
 
 [ ! -f "$OVMF_VARS" ] && cp /usr/share/OVMF/x64/OVMF_VARS.4m.fd "$OVMF_VARS"
+
+CLOUDINIT_DIR="${PROJECT_ROOT}/cloudinit/vms"
+CLOUDINIT_ISO="${CLOUDINIT_DIR}/${VM_NAME}-cloud-init.iso"
 
 qemu-system-x86_64 \
     -enable-kvm \
@@ -27,6 +43,9 @@ qemu-system-x86_64 \
     -blockdev driver=file,filename=${VM_IMG},node-name=drive0_file \
     -blockdev driver=qcow2,file=drive0_file,node-name=drive0_qcow2 \
     -device virtio-blk-pci,drive=drive0_qcow2,bootindex=1 \
+    -blockdev driver=file,filename=${CLOUDINIT_ISO},read-only=on,node-name=ci_file \
+    -blockdev driver=raw,file=ci_file,node-name=ci_raw \
+    -device ide-cd,drive=ci_raw \
     -netdev tap,id=nic${VM_ID},ifname=${TAP_IF},script=no,downscript=no \
-    -device virtio-net,netdev=nic${VM_ID},mac="${VM_MAC}" \
-    -daemonize -display none
+    -device virtio-net,netdev=nic${VM_ID},mac="${VM_MAC}"
+#-daemonize -display none
